Fix MCP refresh notification boundaries
